==> 01_packages_and_introduction.ipynb <==
## ArXiv RAG Search System Using FAISS

## Resources to learn RAG basics

Here's a list of resources that will help you understand what's happening in this notebook:

- Faiss - Introduction to Similarity Search<br>
James Briggs<br>
https://www.youtube.com/watch?v=sKyvsdEv6rk

- Large Language Models with Semantic Search<br>
Deeplearning.Ai short course<br>
https://www.deeplearning.ai/short-courses/large-language-models-semantic-search/

- Colab Notebook that explains rerank<br>
retrieve_rerank_simple_wikipedia.ipynb<br>
https://colab.research.google.com/github/UKPLab/sentence-transformers/blob/master/examples/applications/retrieve_rerank/retrieve_rerank_simple_wikipedia.ipynb#scrollTo=UlArb7kqN3Re

- Vector Databases: from Embeddings to Applications<br>
Deeplearning.Ai short course<br>
(This approach is an alternative to using FAISS and Sentence Transformers)<br>
https://www.deeplearning.ai/short-courses/vector-databases-embeddings-applications/

- Sentence transformers docs<br>
https://www.sbert.net/



## Objectives

- Build a RAG (Retrieval Augmented Generation) search system that uses vector search to compare natural language search queries to research paper titles and abstracts in the ArXiv database.
- Use free tools like Sentence Transformers to create the vectors and FAISS to manage the vector search.
- Format the search results to enable quick review.
- Use OpenAi to create a natural language ouput.

##### Example search query: "I want to build an invisibility cloak like the one in Harry Potter."

## Approach

Arxiv has a dataset on Kaggle that gets updated weekly. It includes a file containing metadata for all papers stored in the Arxiv database. The metadata includes the title and abstract of each paper. There are approximately 2.4 million papers in the ArXiv dataset.

To create the RAG system we will take the title and abstract for each paper and convert it into a vector embedding. These vectors will be stored in a FAISS index.

FAISS (Facebook AI Similarity Search) is an open-source library designed for fast (GPU supported) vector similarity search in large datasets.

When a search query is submitted, the query text string will be vectorized. This query vector will then be compared to the vectors in the FAISS index. Then, the search results will be reranked and the top matches will be returned. The paper title, arxiv categories, paper abstract and a link to the pdf file will be displayed for each search result. This format will enable users to quickly scan through the search results to determine which papers are relevant to their work.

Finally, an OpenAi model wil be used to create a one sentence summary of each abstract in the search results.

## Install Packages

!pip install -U sentence-transformers

!pip install faiss-cpu
#!pip install faiss-gpu

!pip install openai

!pip install rank-bm25

import pandas as pd
import numpy as np
import os

import json
import re

# Optional
# Please add your OpenAi API Key here.
# I will delete this key after committing this notebook.
#OPENAI_API_KEY  = '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'

os.listdir('./arxiv')

# All Arxiv category codes
# Source: https://www.kaggle.com/code/artgor/arxiv-metadata-exploration

# https://arxiv.org/category_taxonomy
# https://info.arxiv.org/help/api/user-manual.html#subject_classifications


category_map = {
# These created errors when mapping categories to descriptions
'acc-phys': 'Accelerator Physics',
'adap-org': 'Not available',
'q-bio': 'Not available',
'cond-mat': 'Not available',
'chao-dyn': 'Not available',
'patt-sol': 'Not available',
'dg-ga': 'Not available',
'solv-int': 'Not available',
'bayes-an': 'Not available',
'comp-gas': 'Not available',
'alg-geom': 'Not available',
'funct-an': 'Not available',
'q-alg': 'Not available',
'ao-sci': 'Not available',
'atom-ph': 'Atomic Physics',
'chem-ph': 'Chemical Physics',
'plasm-ph': 'Plasma Physics',
'mtrl-th': 'Not available',
'cmp-lg': 'Not available',
'supr-con': 'Not available',
###

# Added
'econ.GN': 'General Economics',
'econ.TH': 'Theoretical Economics',
'eess.SY': 'Systems and Control',

'astro-ph': 'Astrophysics',
'astro-ph.CO': 'Cosmology and Nongalactic Astrophysics',
'astro-ph.EP': 'Earth and Planetary Astrophysics',
'astro-ph.GA': 'Astrophysics of Galaxies',
'astro-ph.HE': 'High Energy Astrophysical Phenomena',
'astro-ph.IM': 'Instrumentation and Methods for Astrophysics',
'astro-ph.SR': 'Solar and Stellar Astrophysics',
'cond-mat.dis-nn': 'Disordered Systems and Neural Networks',
'cond-mat.mes-hall': 'Mesoscale and Nanoscale Physics',
'cond-mat.mtrl-sci': 'Materials Science',
'cond-mat.other': 'Other Condensed Matter',
'cond-mat.quant-gas': 'Quantum Gases',
'cond-mat.soft': 'Soft Condensed Matter',
'cond-mat.stat-mech': 'Statistical Mechanics',
'cond-mat.str-el': 'Strongly Correlated Electrons',
'cond-mat.supr-con': 'Superconductivity',
'cs.AI': 'Artificial Intelligence',
'cs.AR': 'Hardware Architecture',
'cs.CC': 'Computational Complexity',
'cs.CE': 'Computational Engineering, Finance, and Science',
'cs.CG': 'Computational Geometry',
'cs.CL': 'Computation and Language',
'cs.CR': 'Cryptography and Security',
'cs.CV': 'Computer Vision and Pattern Recognition',
'cs.CY': 'Computers and Society',
'cs.DB': 'Databases',
'cs.DC': 'Distributed, Parallel, and Cluster Computing',
'cs.DL': 'Digital Libraries',
'cs.DM': 'Discrete Mathematics',
'cs.DS': 'Data Structures and Algorithms',
'cs.ET': 'Emerging Technologies',
'cs.FL': 'Formal Languages and Automata Theory',
'cs.GL': 'General Literature',
'cs.GR': 'Graphics',
'cs.GT': 'Computer Science and Game Theory',
'cs.HC': 'Human-Computer Interaction',
'cs.IR': 'Information Retrieval',
'cs.IT': 'Information Theory',
'cs.LG': 'Machine Learning',
'cs.LO': 'Logic in Computer Science',
'cs.MA': 'Multiagent Systems',
'cs.MM': 'Multimedia',
'cs.MS': 'Mathematical Software',
'cs.NA': 'Numerical Analysis',
'cs.NE': 'Neural and Evolutionary Computing',
'cs.NI': 'Networking and Internet Architecture',
'cs.OH': 'Other Computer Science',
'cs.OS': 'Operating Systems',
'cs.PF': 'Performance',
'cs.PL': 'Programming Languages',
'cs.RO': 'Robotics',
'cs.SC': 'Symbolic Computation',
'cs.SD': 'Sound',
'cs.SE': 'Software Engineering',
'cs.SI': 'Social and Information Networks',
'cs.SY': 'Systems and Control',
'econ.EM': 'Econometrics',
'eess.AS': 'Audio and Speech Processing',
'eess.IV': 'Image and Video Processing',
'eess.SP': 'Signal Processing',
'gr-qc': 'General Relativity and Quantum Cosmology',
'hep-ex': 'High Energy Physics - Experiment',
'hep-lat': 'High Energy Physics - Lattice',
'hep-ph': 'High Energy Physics - Phenomenology',
'hep-th': 'High Energy Physics - Theory',
'math.AC': 'Commutative Algebra',
'math.AG': 'Algebraic Geometry',
'math.AP': 'Analysis of PDEs',
'math.AT': 'Algebraic Topology',
'math.CA': 'Classical Analysis and ODEs',
'math.CO': 'Combinatorics',
'math.CT': 'Category Theory',
'math.CV': 'Complex Variables',
'math.DG': 'Differential Geometry',
'math.DS': 'Dynamical Systems',
'math.FA': 'Functional Analysis',
'math.GM': 'General Mathematics',
'math.GN': 'General Topology',
'math.GR': 'Group Theory',
'math.GT': 'Geometric Topology',
'math.HO': 'History and Overview',
'math.IT': 'Information Theory',
'math.KT': 'K-Theory and Homology',
'math.LO': 'Logic',
'math.MG': 'Metric Geometry',
'math.MP': 'Mathematical Physics',
'math.NA': 'Numerical Analysis',
'math.NT': 'Number Theory',
'math.OA': 'Operator Algebras',
'math.OC': 'Optimization and Control',
'math.PR': 'Probability',
'math.QA': 'Quantum Algebra',
'math.RA': 'Rings and Algebras',
'math.RT': 'Representation Theory',
'math.SG': 'Symplectic Geometry',
'math.SP': 'Spectral Theory',
'math.ST': 'Statistics Theory',
'math-ph': 'Mathematical Physics',
'nlin.AO': 'Adaptation and Self-Organizing Systems',
'nlin.CD': 'Chaotic Dynamics',
'nlin.CG': 'Cellular Automata and Lattice Gases',
'nlin.PS': 'Pattern Formation and Solitons',
'nlin.SI': 'Exactly Solvable and Integrable Systems',
'nucl-ex': 'Nuclear Experiment',
'nucl-th': 'Nuclear Theory',
'physics.acc-ph': 'Accelerator Physics',
'physics.ao-ph': 'Atmospheric and Oceanic Physics',
'physics.app-ph': 'Applied Physics',
'physics.atm-clus': 'Atomic and Molecular Clusters',
'physics.atom-ph': 'Atomic Physics',
'physics.bio-ph': 'Biological Physics',
'physics.chem-ph': 'Chemical Physics',
'physics.class-ph': 'Classical Physics',
'physics.comp-ph': 'Computational Physics',
'physics.data-an': 'Data Analysis, Statistics and Probability',
'physics.ed-ph': 'Physics Education',
'physics.flu-dyn': 'Fluid Dynamics',
'physics.gen-ph': 'General Physics',
'physics.geo-ph': 'Geophysics',
'physics.hist-ph': 'History and Philosophy of Physics',
'physics.ins-det': 'Instrumentation and Detectors',
'physics.med-ph': 'Medical Physics',
'physics.optics': 'Optics',
'physics.plasm-ph': 'Plasma Physics',
'physics.pop-ph': 'Popular Physics',
'physics.soc-ph': 'Physics and Society',
'physics.space-ph': 'Space Physics',
'q-bio.BM': 'Biomolecules',
'q-bio.CB': 'Cell Behavior',
'q-bio.GN': 'Genomics',
'q-bio.MN': 'Molecular Networks',
'q-bio.NC': 'Neurons and Cognition',
'q-bio.OT': 'Other Quantitative Biology',
'q-bio.PE': 'Populations and Evolution',
'q-bio.QM': 'Quantitative Methods',
'q-bio.SC': 'Subcellular Processes',
'q-bio.TO': 'Tissues and Organs',
'q-fin.CP': 'Computational Finance',
'q-fin.EC': 'Economics',
'q-fin.GN': 'General Finance',
'q-fin.MF': 'Mathematical Finance',
'q-fin.PM': 'Portfolio Management',
'q-fin.PR': 'Pricing of Securities',
'q-fin.RM': 'Risk Management',
'q-fin.ST': 'Statistical Finance',
'q-fin.TR': 'Trading and Market Microstructure',
'quant-ph': 'Quantum Physics',
'stat.AP': 'Applications',
'stat.CO': 'Computation',
'stat.ME': 'Methodology',
'stat.ML': 'Machine Learning',
'stat.OT': 'Other Statistics',
'stat.TH': 'Statistics Theory'
}

# https://www.kaggle.com/code/matthewmaddock/nlp-arxiv-dataset-transformers-and-umap

# This takes about 1 minute.


cols = ['id', 'title', 'abstract', 'categories']
data = []
file_name = './arxiv/arxiv-metadata-oai-snapshot.json'


with open(file_name, encoding='latin-1') as f:
    for line in f:
        doc = json.loads(line)
        lst = [doc['id'], doc['title'], doc['abstract'], doc['categories']]
        data.append(lst)

df_data = pd.DataFrame(data=data, columns=cols)

print(df_data.shape)

df_data.head()

## Convert the category codes into text

If a description was not available for a category code then I left it out when converting category codes into text.

def get_cat_text(x):

    cat_text = ''

    # Put the codes into a list
    cat_list = x.split(' ')

    for i, item in enumerate(cat_list):

        cat_name = category_map[item]

        # If there was no description available
        # for the category code then don't include it in the text.
        if cat_name != 'Not available':

            if i == 0:
                cat_text = cat_name
            else:
                cat_text = cat_text + ', ' + cat_name

    # Remove leading and trailing spaces
    cat_text = cat_text.strip()

    return cat_text


df_data['cat_text'] = df_data['categories'].apply(get_cat_text)

df_data.head()

# Print details of one paper

i = 1

print('Id:',df_data.loc[i, 'id'])
print()
print('Title:',df_data.loc[i, 'title'])
print()
print('Categories:',df_data.loc[i, 'cat_text'])
print()
print('Abstract:',df_data.loc[i, 'abstract'])

## Clean the text
- Replace newline characters ('\n') with a space
- Remove leading and trailing spaces

# Replace newline characters ('\n') with a space
# Remove leading and trailing spaces

def clean_text(x):

    # Replace newline characters with a space
    new_text = x.replace("\n", " ")
    # Remove leading and trailing spaces
    new_text = new_text.strip()

    return new_text

df_data['title'] = df_data['title'].apply(clean_text)
df_data['abstract'] = df_data['abstract'].apply(clean_text)

#df_filtered.head()

## Create the text string that will be vectorized

Here the title will be appended to the abstract.

# Append the title to the abstract

df_data['prepared_text'] = df_data['title'] + ' {title} ' + df_data['abstract']

#df_data.head()

## Get the data ready for vectorizing

We need a list of text strings.

# Create a list of text chunks

chunk_list = list(df_data['prepared_text'])

# The ids are used to create web links to each paper.
# You can access each paper directly on ArXiv using these links:
# https://arxiv.org/abs/{id}: ArXiv page for the paper
# https://arxiv.org/pdf/{id}: Direct link to download the PDF

arxiv_id_list = list(df_data['id'])
cat_list = list(df_data['cat_text'])

print(len(chunk_list))
print(len(arxiv_id_list))
print(len(cat_list))

chunk_list[0]

## Create the embedding vetors

This step takes about 1 hour 40 minutes to create approx. 2.4 million vectors. I saved these vectors and created a Part 2 notebook where the saved vectors are loaded instead of being created from scratch.

You can use the Part 2 notebook to quickly test this RAG system. There you can enter your own search queries and review the results.

from sentence_transformers import SentenceTransformer

model = SentenceTransformer("all-MiniLM-L6-v2")

# Sentences are encoded by calling model.encode()
embeddings = model.encode(chunk_list)

print(embeddings.shape)
print('Embedding length', embeddings.shape[1])

# Display one embedding

i = 1
print(chunk_list[i])
print(embeddings[i])

## Save the embedding vectors and the dataframe

type(embeddings)


# Save the array in compressed format
np.savez_compressed('compressed_array.npz', array_data=embeddings)

!ls

# Check the size of the saved file

import os

# Get the size of the file in bytes
file_size_bytes = os.path.getsize('compressed_array.npz')

# Convert bytes to megabytes
file_size_mb = file_size_bytes / (1024 * 1024)

print("File size:", file_size_mb, "MB")

# How to load the saved array

# Load the compressed array
loaded_embeddings = np.load('compressed_array.npz')

# Access the array by the name you specified ('my_array' in this case)
loaded_embeddings = loaded_embeddings['array_data']

loaded_embeddings.shape

# Save the DataFrame in compressed format

df_data.to_csv('compressed_dataframe.csv.gz', compression='gzip', index=False)

!ls

df = pd.read_csv('compressed_dataframe.csv.gz', compression='gzip')

print(df.shape)

df.head(2)

## How to set up FAISS for Exhaustive Search

In an exhaustive search (brute-force search) we compare the query vector to every vector in the database. Therefore, we don't need to train an index.

You'll need to have watched this video to understand all that we are going to do next:<br>
Faiss - Introduction to Similarity Search<br>
https://www.youtube.com/watch?v=sKyvsdEv6rk

import faiss

embed_length = embeddings.shape[1]

index = faiss.IndexFlatL2(embed_length)

# Check if the index is trained.
# No training needed when using greedy search i.e. IndexFlatL2
index.is_trained

# Add the embeddings to the index

index.add(embeddings)

# Check the total number of embeddings in the index
index.ntotal

# Run a query

query_text = """
I want to quantify the data that can be reconstructed from gradients of trained model
"""
query = [query_text]


# Vectorize the query string
query_embedding = model.encode(query)

# Set the number of outputs we want
top_k = 3

# Run the query
# index_vals refers to the chunk_list index values
scores, index_vals = index.search(query_embedding, top_k)

print(index_vals)
print(scores)

# Let's print the first search result

pred_indexes = index_vals[0]

i = 0
chunk_index = pred_indexes[i]
text = chunk_list[chunk_index]

text

## Set up FAISS - Nearest Neighbor Search

Exhaustive (brute-force) search can be slow when searching over a large number of vectors. A nearest neighbor search is faster.

# How many clusters (voronoid cells) do we want?
# Example: For 4 centroilds we need at least 156 embeddings in
# order to train the index.
num_centroids = 5

quantizer = faiss.IndexFlatL2(embed_length)

index = faiss.IndexIVFFlat(quantizer, embed_length, num_centroids)

# Train the index
# After the index is trained it's ready to receive data

index.train(embeddings)

index.is_trained

# Add the embeddings to the index

index.add(embeddings)

# Check how many embeddings are in the index
index.ntotal

query = [query_text]
query_embedding = model.encode(query)

top_k = 5


# Run the query
# index_vals refers to the chunk_list index values
scores, index_vals = index.search(query_embedding, top_k)

print(index_vals)
print(scores)

# Let's print the first search result

pred_indexes = index_vals[0]

i = 3
chunk_index = pred_indexes[i]
text = chunk_list[chunk_index]
